+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
+* [#24](https://github.com/skuzzle/snapshot-tests/issues/24): Only create snapshot directory if necessary
 * Simple API to specify custom rules using XPath for XML comparison
 * Change: Custom json matchers are only tested against the new actual result and not against the persisted snapshot anymore
-* [#24](https://github.com/skuzzle/snapshot-tests/issues/24): Only create snapshot directory if necessary
+* Add convenience method `TestFile.asText()` which defaults to use UTF-8 as file encoding
 
 ### Experimental
 * Fix: object traversal failed on primitive arrays